requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.7.3
psycopg2==2.7.3.2
